python-ads-service/requirements.txt
Flask==3.0.0
google-ads==28.0.0
python-dotenv==1.0.0
